requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.2.0
pyyaml>=6.0
